+++ packages/mui-utils/src/getReactElementRef/index.ts
@@ -1,0 +1,1 @@
+export { default } from './getReactElementRef';
