--- packages/mui-utils/src/getReactNodeRef/index.ts
@@ -1,1 +1,0 @@
-export { default } from './getReactNodeRef';
